requirements.txt
docker==7.0.0
click==8.1.7
pytest==7.4.3
black==23.12.1
flake8==7.0.0
tabulate==0.9.0
pyyaml==6.0.1 